+++ jobbergate-cli/tests/test_auth.py
@@ -96,6 +96,16 @@ def test_validate_token_and_extract_identity__re_raises_ExpiredSignatureError(ma
             validate_token_and_extract_identity(TokenSet(access_token=access_token))
 
 
+def test_validate_token_and_extract_identity__raises_abort_on_empty_token():
+    """
+    Validate that the ``validate_token_and_extract_identity()`` function will
+    raise an ``Abort`` when the access_token exists but is an empty string/file.
+    """
+    test_token_set = TokenSet(access_token="")
+    with pytest.raises(Abort, match="Access token file exists but it is empty"):
+        validate_token_and_extract_identity(test_token_set)
+
+
 def test_validate_token_and_extract_identity__raises_abort_on_unknown_error(mocker):
     """
     Validate that the ``validate_token_and_extract_identity()`` function will raise an ``Abort`` when the